requirements.txt
asyncio==3.4.3
discord.py==1.7.3
aiofiles==0.8.0
aiomysql==0.0.22
Pillow==8.4.0
psutil==5.8.0
filetype==1.0.8
aiohttp==3.7.4
ksoftapi==0.4.1
html5lib==1.1
beautifulsoup4==4.10.0
dislash.py==1.4.9
